updated comparison script

+++ performance_comparison.py
@@ -4,6 +4,7 @@ import shlex
 import numpy as np 
 import matplotlib.pyplot as plt
 import shutil
+import os
   
 def wordcount(cmd):
     # Execute WordCount 3 times and compute the average execution time
@@ -13,22 +14,22 @@ def wordcount(cmd):
         output = run(shlex.split(cmd), stdout = PIPE, stderr = PIPE, universal_newlines=True).stderr
         # get user time
         for line in output.splitlines():
-            if "user" in line:
+            if "user " in line:
                 user_index = line.find("user")
                 user_time = float(line[:user_index])
                 print("user_time for execution ", i + 1 , ": ", user_time)
                 total_execution_time += user_time
-	# remove output directory
-        try:
-            shutil.rmtree(output_directory)
-        except OSError as e:
-            print("Error: %s - %s." % (e.filename, e.strerror))
+        if os.path.isdir(output_directory):
+            try:
+                shutil.rmtree(output_directory)
+            except OSError as e:
+                print("Error: %s - %s." % (e.filename, e.strerror))
     average_time = round(total_execution_time / 3, 2)
     print("average time: ", average_time)
     return average_time
 
 def plot(hadoop, spark):
-    X = ['Dataset1', 'Dataset2', 'Dataset3', 'Dataset4', 'Dataset5', 'Dataset6', 'Dataset7', 'Dataset8', 'Dataset9']
+    X = ['1', '2', '3', '4', '5', '6', '7', '8', '9']
     
     X_axis = np.arange(len(X))
 
@@ -44,11 +45,18 @@ def plot(hadoop, spark):
   
 if __name__ == '__main__':
 
-    filename = 'dataset1.txt'
+    hadoop_execution_times = []
+    spark_execution_times = []
     output_directory = '/home/azureuser/datasets/output'
-
-    hadoop_cmd = 'time hadoop jar /usr/local/hadoop-3.3.1/share/hadoop/mapreduce/hadoop-mapreduce-examples-3.3.1.jar wordcount /home/azureuser/datasets/' + filename + ' ' + output_directory
-    spark_cmd = 'time python3 /home/azureuser/wordcount.py /home/azureuser/datasets/' + filename + ' ' + output_directory
-
-    # wordcount(hadoop_cmd)
-    wordcount(spark_cmd)
+    
+    for i in range(1, 10):
+        filename = 'dataset' + str(i) + '.txt'
+        hadoop_cmd = 'time python3 /home/azureuser/wordcount_hadoop.py -r local /home/azureuser/datasets/' + filename
+        spark_cmd = 'time python3 /home/azureuser/wordcount_spark.py /home/azureuser/datasets/' + filename + ' ' + output_directory
+        print("starting hadoop wordcount on", filename)
+        hadoop_execution_times.append(wordcount(hadoop_cmd))
+        print("starting spark wordcount on", filename)
+        spark_execution_times.append(wordcount(spark_cmd))
+    print("Hadoop WordCount execution times: ", hadoop_execution_times)
+    print("Spark WordCount execution times: ", spark_execution_times)
+    plot(hadoop_execution_times, spark_execution_times)
